+++ examples/minicpmv/CMakeLists.txt
@@ -33,10 +33,10 @@ endif()
 set(TARGET minicpmv-cli)
 add_executable(minicpmv-cli minicpmv-cli.cpp)
 install(TARGETS minicpmv-cli RUNTIME)
-target_link_libraries(minicpmv-cli PRIVATE common minicpmv_io minicpmv ${CMAKE_THREAD_LIBS_INIT})
+target_link_libraries(minicpmv-cli PRIVATE common minicpmv_wrapper minicpmv ${CMAKE_THREAD_LIBS_INIT})
 target_compile_features(minicpmv PRIVATE cxx_std_11)
 
-add_library(minicpmv_io OBJECT
-        minicpmv_io.cpp
+add_library(minicpmv_wrapper OBJECT
+        minicpmv_wrapper.cpp
 )
-target_link_libraries(minicpmv_io PRIVATE minicpmv ${CMAKE_THREAD_LIBS_INIT})
+target_link_libraries(minicpmv_wrapper PRIVATE minicpmv ${CMAKE_THREAD_LIBS_INIT})
