+++ examples/minicpmv/minicpmv_wrapper.h
